fix: refactor observation
